fix CI by droping .toml

+++ examples/janus/inference.py
@@ -65,9 +65,9 @@ def multimodal_understanding(
     tokenizer = vl_chat_processor.tokenizer
 
     pil_images = load_pil_images(conversation)
-    prepare_inputs = vl_chat_processor(
-        conversations=conversation, images=pil_images, force_batchify=True
-    ).to(vl_gpt.dtype)
+    prepare_inputs = vl_chat_processor(conversations=conversation, images=pil_images, force_batchify=True).to(
+        vl_gpt.dtype
+    )
 
     inputs_embeds = vl_gpt.prepare_inputs_embeds(**prepare_inputs)
     st = time()
@@ -102,15 +102,9 @@ def multimodal_understanding(
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--ms_mode", type=int, default=1, help="mindspore mode, 0: graph, 1: pynative"
-    )
-    parser.add_argument(
-        "--image", type=str, default="images/doge.png", help="path to input image"
-    )
-    parser.add_argument(
-        "--question", type=str, default="explain this meme", help="path to input image"
-    )
+    parser.add_argument("--ms_mode", type=int, default=1, help="mindspore mode, 0: graph, 1: pynative")
+    parser.add_argument("--image", type=str, default="images/doge.png", help="path to input image")
+    parser.add_argument("--question", type=str, default="explain this meme", help="path to input image")
     parser.add_argument(
         "--ckpt_path",
         type=str,
@@ -145,9 +139,7 @@ if __name__ == "__main__":
         ms.set_context(jit_config={"jit_level": "O0"})
 
     # specify the path to the model
-    vl_chat_processor: VLChatProcessor = VLChatProcessor.from_pretrained(
-        args.model_path
-    )
+    vl_chat_processor: VLChatProcessor = VLChatProcessor.from_pretrained(args.model_path)
 
     config = AutoConfig.from_pretrained(args.model_path)
     language_config = config.language_config
@@ -160,16 +152,12 @@ if __name__ == "__main__":
         vl_gpt = set_model_param_dtype(vl_gpt, dtype)
 
         parameter_dict = ms.load_checkpoint(args.ckpt_path)
-        param_not_load, ckpt_not_load = ms.load_param_into_net(
-            vl_gpt, parameter_dict, strict_load=True
-        )
+        param_not_load, ckpt_not_load = ms.load_param_into_net(vl_gpt, parameter_dict, strict_load=True)
         print("net param not load: {}".format(param_not_load))
         print("ckpt param not load: {}".format(ckpt_not_load))
     else:
         with no_init_parameters():
-            vl_gpt = MultiModalityCausalLM.from_pretrained(
-                args.model_path, config=config
-            )
+            vl_gpt = MultiModalityCausalLM.from_pretrained(args.model_path, config=config)
         dtype = ms.bfloat16
         vl_gpt = set_model_param_dtype(vl_gpt, dtype)
